==> 1b_get_glove_embeddings.ipynb <==
# !pip install transformers emoji nltk

# import packages
import pickle
import os
import json
import numpy as np
import pandas as pd
import re
import time
from multiprocessing import Pool

GLOVE_DIR = '../../glove'
Glove_path = os.path.join(GLOVE_DIR, 'glove.6B.300d.txt')
print(Glove_path)

embeddings_index = dict()
f = open(Glove_path, encoding="utf8")
for line in f:
    values = line.split()
    word = values[0]
    try:
        coefs = np.asarray(values[1:], dtype='float32')
    except:
        print("Warnning"+str(values)+" in" + str(line))
    embeddings_index[word] = coefs
f.close()
print('Total %s word vectors.' % len(embeddings_index))

dataset_name = 'wos46985'
base_dir = '../../data/WOS/'
data_file = base_dir+'Meta-data/Data.csv'

df = pd.read_csv(data_file)
print('Num Data:',len(df))

df.head(2)

def clean_str(string):
    string = re.sub(r"[^A-Za-z0-9(),!?\'\`]", " ", string)
    string = re.sub(r"\'s", " \'s", string)
    string = re.sub(r"\'ve", " \'ve", string)
    string = re.sub(r"n\'t", " n\'t", string)
    string = re.sub(r"\'re", " \'re", string)
    string = re.sub(r"\'d", " \'d", string)
    string = re.sub(r"\'ll", " \'ll", string)
    string = re.sub(r",", " , ", string)
    string = re.sub(r"!", " ! ", string)
    string = re.sub(r"\(", " \( ", string)
    string = re.sub(r"\)", " \) ", string)
    string = re.sub(r"\?", " \? ", string)
    string = re.sub(r"\s{2,}", " ", string)
    return string.strip().lower()

def get_glove_embeddings(texts, method = 'pool'):
    output_reps_pool = []

    for step,text in enumerate(texts):
        # progress update after every 1000 texts.
        if step % 1000 == 0 and not step == 0:
            print('  Batch {:>5,}  of  {:>5,}.'.format(step, len(texts)))

        base_output = []
        for word in text.split():
            try:
                base_output.append(embeddings_index[word])
            except:
                pass
        
        if len(base_output)>0:
            if method=='pool':
                # Mean Pool
                output_representation = np.mean(base_output, axis=0)
            elif method=='cls':
                output_representation = base_output[0]
            output_reps_pool.append(output_representation)
        else:
            output_reps_pool.append(np.zeros(embeddings_index['a'].shape))
        
    output_reps_df = np.vstack(output_reps_pool)
    print('Embeddings Shape:',output_reps_df.shape)
    
    return output_reps_df

method = 'pool'
text_embedding_dir = '../data/'+dataset_name

texts = df.Abstract.apply(clean_str).tolist()

text_embeddings = get_glove_embeddings(texts, method)
text_embedding_file = text_embedding_dir+'/glove300-embedding-'+method+'.pkl'

if text_embeddings.shape[0] == len(texts):
    print('Saving Embeddings...')
    pickle.dump(text_embeddings, open(text_embedding_file, "wb"))
