==> 2a_get_hierarchy.ipynb <==
import pickle
import os
import json
import numpy as np
import pandas as pd
import re

method = 'pool'

dataset_name = 'wos46985'
base_dir = '../../data/WOS/Meta-data/'
taxonomy_file = base_dir+'wos.taxnomy'

text_embedding_dir = '../data/wos46985'
taxonomy_3_level_info_file = text_embedding_dir+'/taxonomy_3_level_info_file.json'

tx_child_dict = dict(
    (line.strip().split('\t')[0],line.strip().split('\t')[1:]) 
    for line in open(taxonomy_file).readlines()
)
tx_parent_dict = dict()

level_0 = 'Root'
level_1 = tx_child_dict[level_0]
for node in level_1:
    tx_parent_dict[node] = level_0
    
level_2 = []
for node in level_1:
    try:
        level_2.extend(tx_child_dict[node])
        for cnode in tx_child_dict[node]:
            tx_parent_dict[cnode] = node
    except:
        pass

## Stop at level 3

taxonomy_3_level_info = {
    'level_0' : level_0,
    'level_1' : level_1,
    'level_2' : level_2,
    'taxonomy_child_map' : tx_child_dict,
    'taxonomy_parent_map' : tx_parent_dict
}

with open(taxonomy_3_level_info_file, "w") as outfile:
    json.dump(taxonomy_3_level_info, outfile, indent = 4)
