==> 3_get_seed_embeddings.ipynb <==
import json
import pickle
import numpy as np
import pandas as pd

method = 'pool'
embedding_type = 'roberta' # roberta, glove300, fastText

dataset_name = 'wos46985'
base_dir = '../../data/WOS/'
data_file = base_dir+'Meta-data/Data.csv'

text_embedding_dir = '../data/'+dataset_name
text_embedding_file = text_embedding_dir+'/'+embedding_type+'-embedding-'+method+'.pkl'

seed_indices_file = text_embedding_dir+'/seed_indices.json'
remaining_indices_file = text_embedding_dir+'/remaining_indices.json'

seed_embeddings_file = text_embedding_dir+'/cluster_hierarchy_seed_embeddings-roberta-'+method+'.pkl'

text_embeddings = pickle.load(open(text_embedding_file, "rb"))
print(text_embeddings.shape)

with open(seed_indices_file, "r") as infile:
    seeds = json.load(infile)

def get_topic_seeds(seedTopics, topic_seeds={}, main_topic_update_method=1, main_topic_weight=3):
    # Function to retrieve BERT embeddings from memory
    def get_embedding_text(index):
        topic_embedding = text_embeddings[index]
        return topic_embedding

    if include_sub_topics == True:
        main_topics = list(seedTopics.keys())
        topics = list(seedTopics.keys())

        for topic in topics:
            if topic not in topic_seeds.keys():
                indices = seedTopics[topic]['indices']
                topic_embedding = np.mean([get_embedding_text(ind) for ind in indices], axis=0)
                topic_seeds[topic] = topic_embedding

            try:
                for sub_topic in list(seedTopics[topic]['sub-topic'].keys()):
                    if topic+'/'+sub_topic not in topic_seeds.keys():
                        indices = seedTopics[topic]['sub-topic'][sub_topic]['indices']
                        topic_embedding = np.mean([get_embedding_text(ind) for ind in indices], axis=0)
                        topic_seeds[topic+'/'+sub_topic] = topic_embedding

                # Update topic embedding by taking mean of itself with all sub-topic embeddings
                if main_topic_update_method == 1:
                    topic_seeds[topic] = np.mean(
                        [
                          topic_seeds[topic+'/'+sub_topic] for sub_topic in list(seedTopics[topic]['sub-topic'].keys())
                        ] + [topic_seeds[topic]]*main_topic_weight
                        , axis=0
                    )
                elif main_topic_update_method == 2:
                    topic_seeds[topic] = np.mean(
                        [
                            np.mean(
                                [
                                  topic_seeds[topic+'/'+sub_topic] for sub_topic in list(seedTopics[topic]['sub-topic'].keys())
                                ]
                                , axis=0
                            )
                        ]
                        + [topic_seeds[topic]]*main_topic_weight
                        , axis=0
                    )
                elif main_topic_update_method == 3:
                    topic_seeds[topic] = np.mean(
                        [
                            np.mean(
                                [
                                  topic_seeds[topic+'/'+sub_topic] for sub_topic in list(seedTopics[topic]['sub-topic'].keys())
                                ]
                                , axis=0
                            )
                        ]
                        + [topic_seeds[topic]]*main_topic_weight*len(list(seedTopics[topic]['sub-topic'].keys()))
                        , axis=0
                    )
                else:
                    print('Not Altering Main Topic')
            except:
                print(topic,'Whoohs!')
                pass
    else:
        topics = list(seedTopics.keys())

        for topic in topics:
            if topic not in topic_seeds.keys():
                indices = seedTopics[topic]['indices']
                topic_embedding = np.mean([get_embedding_text(ind) for ind in indices], axis=0)
                topic_seeds[topic]= topic_embedding
    
    return topic_seeds

include_sub_topics = True
method = 'pool'

main_topic_update_method=2
main_topic_weight=2

topic_seeds = get_topic_seeds(
    seeds, {}, main_topic_update_method=main_topic_update_method, main_topic_weight=main_topic_weight
)

print(len(list(topic_seeds.keys())),'seeds:',list(topic_seeds.keys()),'\n')

pickle.dump(topic_seeds, open(seed_embeddings_file, "wb"))
